+++ large_image/cache_util/cachefactory.py
@@ -24,10 +24,7 @@ try:
 except ImportError:
     psutil = None
 
-try:
-    from importlib.metadata import entry_points
-except ImportError:
-    from importlib_metadata import entry_points
+from importlib.metadata import entry_points
 
 from .. import config
 from ..exceptions import TileCacheError