+++ sources/mapnik/large_image_source_mapnik/__init__.py
@@ -15,6 +15,8 @@
 #############################################################################
 
 import functools
+from importlib.metadata import PackageNotFoundError
+from importlib.metadata import version as _importlib_version
 
 import mapnik
 import PIL.Image
@@ -26,12 +28,6 @@ from large_image.constants import TILE_FORMAT_PIL, SourcePriority
 from large_image.exceptions import TileSourceError
 from large_image.tilesource.utilities import JSONDict
 
-try:
-    from importlib.metadata import PackageNotFoundError
-    from importlib.metadata import version as _importlib_version
-except ImportError:
-    from importlib_metadata import PackageNotFoundError
-    from importlib_metadata import version as _importlib_version
 try:
     __version__ = _importlib_version(__name__)
 except PackageNotFoundError: