+++ sources/tifffile/large_image_source_tifffile/__init__.py
@@ -3,6 +3,8 @@ import logging
 import math
 import os
 import threading
+from importlib.metadata import PackageNotFoundError
+from importlib.metadata import version as _importlib_version
 
 import numpy as np
 import zarr
@@ -15,12 +17,6 @@ from large_image.tilesource import FileTileSource
 
 tifffile = None
 
-try:
-    from importlib.metadata import PackageNotFoundError
-    from importlib.metadata import version as _importlib_version
-except ImportError:
-    from importlib_metadata import PackageNotFoundError
-    from importlib_metadata import version as _importlib_version
 try:
     __version__ = _importlib_version(__name__)
 except PackageNotFoundError: