+++ test/test_converter.py
@@ -1,7 +1,6 @@
 import json
 import os
 import shutil
-import sys
 
 import large_image_converter
 import large_image_converter.__main__ as main
@@ -317,7 +316,6 @@ def testConvertFromTestSourceFrameArray(tmpdir):
     assert len(info['ifds']) == 6
 
 
-@pytest.mark.skipif(sys.version_info < (3, 7), reason='requires python >= 3.7 for the test image')
 def testConvertImageJ(tmpdir):
     imagePath = datastore.fetch('synthetic_imagej.tiff')
     outputPath = os.path.join(tmpdir, 'out.tiff')