+++ test/test_source_multi.py
@@ -1,6 +1,5 @@
 import json
 import os
-import sys
 
 import large_image_source_multi
 import numpy as np
@@ -76,7 +75,6 @@ def testTilesFromMultiSimpleScaling():
         utilities.checkTilesZXY(source, tileMetadata, tileParams={'frame': frame})
 
 
-@pytest.mark.skipif(sys.version_info < (3, 7), reason='requires python >= 3.7 for a sub-source')
 def testTilesFromMultiMultiSource(multiSourceImagePath):
     imagePath = multiSourceImagePath
     source = large_image_source_multi.open(imagePath)
@@ -135,7 +133,6 @@ def testTilesFromNonschemaMultiString():
         large_image_source_multi.open(sourceString)
 
 
-@pytest.mark.skipif(sys.version_info < (3, 7), reason='requires python >= 3.7 for a sub-source')
 def testInternalMetadata(multiSourceImagePath):
     imagePath = multiSourceImagePath
     source = large_image_source_multi.open(imagePath)
@@ -143,7 +140,6 @@ def testInternalMetadata(multiSourceImagePath):
     assert 'frames' in metadata
 
 
-@pytest.mark.skipif(sys.version_info < (3, 7), reason='requires python >= 3.7 for a sub-source')
 def testAssociatedImages(multiSourceImagePath):
     imagePath = multiSourceImagePath
     source = large_image_source_multi.open(imagePath)
@@ -202,7 +198,6 @@ def testFramesAsAxes():
     assert tileMetadata['IndexRange']['IndexZ'] == 8
 
 
-@pytest.mark.skipif(sys.version_info < (3, 7), reason='requires python >= 3.7 for a sub-source')
 def testMultiComposite():
     datastore.fetch('ITGA3Hi_export_crop2.nd2')
     imagePath = datastore.fetch('multi-source-composite.yaml')