+++ docs/make_docs.sh
@@ -35,6 +35,7 @@ sphinx-apidoc -f -o _build/large_image_source_test ../sources/test/large_image_s
 sphinx-apidoc -f -o _build/large_image_source_tiff ../sources/tiff/large_image_source_tiff
 sphinx-apidoc -f -o _build/large_image_source_tifffile ../sources/tifffile/large_image_source_tifffile
 sphinx-apidoc -f -o _build/large_image_source_vips ../sources/vips/large_image_source_vips
+sphinx-apidoc -f -o _build/large_image_source_zarr ../sources/zarr/large_image_source_zarr
 sphinx-apidoc -f -o _build/large_image_converter ../utilities/converter/large_image_converter
 sphinx-apidoc -f -o _build/large_image_tasks ../utilities/tasks/large_image_tasks
 sphinx-apidoc -f -o _build/girder_large_image ../girder/girder_large_image